fix: postgres migration after type change

+++ golem-registry-service/db/migration/postgres/012_component_file_size.sql
@@ -1,1 +1,1 @@
-ALTER TABLE component_files ADD COLUMN file_size BIGINT NOT NULL DEFAULT 0;
+ALTER TABLE component_files ADD COLUMN file_size NUMERIC NOT NULL DEFAULT 0;
